refacto(folios) refactor collect card
