+++ app/controllers/product/card_folios_controller.ts
@@ -5,15 +5,10 @@ import { errors as vineErrors } from '@vinejs/vine'
 import NotFoundException from '#exceptions/not_found_exception'
 import FolioService from '#services/folio_service'
 import { SuccessOutputDTO } from '#types/success_output_dto_type'
-import {
-  collectValidator,
-  occurrenceValidator,
-  removeMainValidator,
-} from '#validators/card_folio_validator'
+import { occurrenceValidator, removeMainValidator } from '#validators/card_folio_validator'
 import CardService from '#services/card_service'
 import CardFolioService from '#services/card_folio_service'
 import ValidationException from '#exceptions/validation_exception'
-import DuplicateEntryException from '#exceptions/duplicate_entry_exception'
 
 @inject()
 export default class CardFoliosController {
@@ -22,30 +17,6 @@ export default class CardFoliosController {
     private cardService: CardService,
     private cardFolioService: CardFolioService
   ) {}
-
-  async collect({ request, response, authUser }: HttpContext) {
-    try {
-      const payload = await collectValidator.validate(request.all())
-      const card = await this.cardService.getCardIdById(payload.cardId)
-      const folio = await this.folioService.getMainFolioByUserId(authUser.id)
-      await this.cardFolioService.createCardMainFolio(card.id, folio.id)
-      const successResponse: SuccessOutputDTO = {
-        message: 'Card added to your main folio successfully',
-      }
-      return response.ok(successResponse)
-    } catch (error) {
-      if (error instanceof vineErrors.E_VALIDATION_ERROR) {
-        throw new ValidationException(error)
-      }
-      if (error.code === 'ER_DUP_ENTRY') {
-        throw new DuplicateEntryException('Card already exists in the folio')
-      }
-      if (error instanceof lucidErrors.E_ROW_NOT_FOUND) {
-        throw new NotFoundException(error)
-      }
-      throw error
-    }
-  }
 
   async occurrence({ request, response, authUser }: HttpContext) {
     try {
