+++ start/routes.ts
@@ -36,7 +36,7 @@ router
             router.get('/:id/cards', [ChildFoliosController, 'cards']).use(middleware.auth())
             router.post('/', [ChildFoliosController, 'store']).use(middleware.auth())
             router.post('/init', [MainFoliosController, 'store']).use(middleware.auth())
-            router.post('/cards', [CardFoliosController, 'collect']).use(middleware.auth())
+            router.post('/cards', [MainFoliosController, 'collect']).use(middleware.auth())
             router.patch('/cards/:id', [CardFoliosController, 'occurrence']).use(middleware.auth())
             router.delete('/cards/:id', [CardFoliosController, 'delete']).use(middleware.auth())
           })
